Forward TERM signal to child processes.
The original buildpack was made for Heroku and was taking in cosideration that signals are forwarded to every process within the DYNO.

On Build.io signals are not sent to all processes within the POD, only to the entry point (PID 1) so that process is responsible of
forwarding signals to subprocess. On this change I'm adding that logic so the "app" sub process can get the signal and terminate
its execution correctly.